requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.126.0
python-telegram-bot==21.1.1
pytube==15.0.0